fix: Group & Progress Components
